requirements.txt
flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
PyYAML
pytz
pdfplumber
scikit-learn
requests
